pyproject.toml
[project]
name = "sitsformer"
version = "0.1.0"
description = "Explore the PyTorch implementation of SITS-Former, a transformer-based model for satellite image time series analysis and classification"
authors = [{name = "Mike Martinez", email = "brillozon@gmail.com"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<4.0"
keywords = ["machine-learning", "pytorch", "transformer", "satellite-imagery", "time-series", "remote-sensing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "torch>=2.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "einops>=0.7.0",
    "pillow>=10.0.0",
    "xarray>=2023.1.0",
    "seaborn>=0.12.0",
    "h5py>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/sitsformer"
Repository = "https://github.com/yourusername/sitsformer"
Documentation = "https://github.com/yourusername/sitsformer"

[project.scripts]
sitsformer-train = "sitsformer.scripts.train:main"
sitsformer-evaluate = "sitsformer.scripts.evaluate:main"
sitsformer-inference = "sitsformer.scripts.inference:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.0",
]
experiment = [
    "wandb>=0.15.0",
    "tensorboard>=2.14.0",
]
geospatial = [
    "rasterio>=1.3.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "fiona>=1.9.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[tool.poetry]
packages = [{include = "sitsformer", from = "src"}]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["sitsformer"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_column_numbers = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "albumentations",
    "rasterio",
    "xarray",
    "einops",
    "sklearn.*",
    "wandb",
    "cv2",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=sitsformer --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/sitsformer"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]